# backend/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: maicivy API
  version: 1.0.0
  description: |
    Interactive CV platform with AI-powered cover letter generation.

    This API provides endpoints for:
    - Adaptive CV generation based on themes
    - AI-powered dual cover letter generation (motivation + anti-motivation)
    - Real-time analytics and visitor tracking
    - GitHub integration for automated CV enrichment
    - Professional timeline and profile detection

    ## Rate Limiting
    - AI letter generation: 5 requests/day with 2-minute cooldown between requests
    - Access gate: Requires 3 visits or detected recruiter/tech profile

    ## Authentication
    Session-based authentication using cookies (session_id).
  contact:
    name: Alexi
    email: contact@maicivy.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.maicivy.com
    description: Production server (if deployed)
...
paths:
...
  /api/v1/cv:
    get:
      tags:
        - CV
      summary: Get adaptive CV
      description: Returns CV adapted to the specified theme with scored experiences and skills
      operationId: getAdaptiveCV
      parameters:
        - name: theme
          in: query
          description: Theme ID to adapt the CV (backend, frontend, fullstack, devops, cpp, artistique)
          required: false
          schema:
            type: string
            enum: [backend, frontend, fullstack, devops, cpp, artistique]
            default: fullstack
      responses:
        '200':
          description: Adaptive CV data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdaptiveCVResponse'
        '400':
          description: Invalid theme
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/experiences:
    get:
      tags:
        - CV
      summary: Get all experiences
      description: Returns all professional experiences
      operationId: getAllExperiences
      responses:
        '200':
          description: List of experiences
          content:
            application/json:
              schema:
                type: object
                properties:
                  experiences:
                    type: array
                    items:
                      $ref: '#/components/schemas/Experience'
                  count:
                    type: integer
        '500':
          description: Failed to fetch experiences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/skills:
    get:
      tags:
        - CV
      summary: Get all skills
      description: Returns all skills with categories and proficiency levels
      operationId: getAllSkills
      responses:
        '200':
          description: List of skills
          content:
            application/json:
              schema:
                type: object
                properties:
                  skills:
                    type: array
                    items:
                      $ref: '#/components/schemas/Skill'
                  count:
                    type: integer
        '500':
          description: Failed to fetch skills
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects:
    get:
      tags:
        - CV
      summary: Get all projects
      description: Returns all projects
      operationId: getAllProjects
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
                  count:
                    type: integer
        '500':
          description: Failed to fetch projects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/cv/export:
    get:
      tags:
        - CV
      summary: Export CV as PDF
      description: Generates and downloads CV as PDF for the specified theme
      operationId: exportCVPDF
      parameters:
        - name: theme
          in: query
          description: Theme ID
          required: false
          schema:
            type: string
            default: fullstack
        - name: format
          in: query
          description: Export format (only PDF supported)
          required: false
          schema:
            type: string
            enum: [pdf]
            default: pdf
      responses:
        '200':
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '400':
          description: Invalid theme or format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to generate PDF
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ====================
  # LETTERS ENDPOINTS
  # ====================
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
        details:
          type: string

    # CV
    AdaptiveCVResponse:
      type: object
      properties:
        theme:
          $ref: '#/components/schemas/CVTheme'
        experiences:
          type: array
          items:
            $ref: '#/components/schemas/ScoredExperience'
        skills:
          type: array
          items:
            $ref: '#/components/schemas/ScoredSkill'
        projects:
          type: array
          items:
            $ref: '#/components/schemas/Project'
        summary:
          type: string
    CVTheme:
      type: object
      properties:
        id:
          type: string
          example: backend
        name:
          type: string
          example: Backend Developer
        description:
          type: string
    Experience:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
          example: Senior Backend Engineer
        company:
          type: string
          example: Google
        description:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
          nullable: true
        technologies:
          type: array
          items:
            type: string
        category:
          type: string
          example: backend
    ScoredExperience:
      allOf:
        - $ref: '#/components/schemas/Experience'
        - type: object
          properties:
            relevance_score:
              type: number
              format: float
              minimum: 0
              maximum: 100
    Skill:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          example: Go
        category:
          type: string
          example: Backend
        proficiency:
          type: integer
          minimum: 1
          maximum: 5
        years_experience:
          type: integer
    ScoredSkill:
      allOf:
        - $ref: '#/components/schemas/Skill'
        - type: object
          properties:
            relevance_score:
              type: number
              format: float
              minimum: 0
              maximum: 100
    Project:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
          example: maicivy
        description:
          type: string
        technologies:
          type: array
          items:
            type: string
        category:
          type: string
        image_url:
          type: string
          format: uri
        github_url:
          type: string
          format: uri
        demo_url:
          type: string
          format: uri
        in_progress:
          type: boolean

    # Letters
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_id
      description: Session-based authentication using cookies
security:
  - sessionCookie: []